*** remove most api public types, add generated protobuf definitions ***
