--- api/version.go
@@ -1,4 +1,0 @@
-package api
-
-// Version API version
-const Version = "v1"
